# src/main/resources/static/openapi.yml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Sorokchat Server
  description: Бекенд для месенджеру SorokChat
  license:
    name: MIT
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен доступу у форматі Bearer <token>
  schemas:
    NewUserPayload:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: Sorokovskys@ukr.net
          description: Електронна адреса.
        password:
          type: string
          format: password
          example: password
          description: Пароль(мінімум 6 символів).
        firstName:
          type: string
          description: Ім'я.
          example: Андрій
        lastName:
          type: string
          description: Прізвище.
          example: Сороковський
        middleName:
          type: string
          description: По батькові.
          example: Іванович
    ErrorPayload:
      type: object
      example:
        email: Invalid format of string for email.
        password: Size of string must be from 6 to 1000000000 characters.
    LoginPayload:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: Електронна адреса.
          example: Sorokovskys@ukr.net
        password:
          type: string
          format: password
          description: Пароль(мінімум 6 символів).
    GetUserPayload:
      properties:
        id:
          type: integer
          format: int64
          description: Унікальний ідентифікатор.
          default: 1
        createdAt:
          type: string
          default: "2025-08-09T18:12:23Z"
          description: Дата створення користувача.
        updatedAt:
          type: string
          default: "2025-08-09T18:12:23Z"
          description: Дата оновлення користувача.
        email:
          type: string
          format: email
          default: Sorokovskys@ukr.net
          description: Електронна адреса.
        password:
          type: string
          format: password
          default: password
          description: Пароль(мінімум 6 символів).
        firstName:
          type: string
          description: Ім'я.
          default: Андрій
        lastName:
          type: string
          description: Прізвище.
          default: Сороковський
        middleName:
          type: string
          description: По батькові.
          default: Іванович
security:
  - bearerAuth: [ ]
paths:
  /authentication/register:
    post:
      tags:
        - Authentication Controller
      summary: Реєстрація користувача.
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUserPayload'
      responses:
        "201":
          description: Користувача успісно зареєстровано.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetUserPayload'
          headers:
            Location:
              description: URL для отримання авторизованого користувача.
              schema:
                type: string
                format: uri
                example: /authentication/get-me
              required: true
            Authorization:
              description: Токен доступу.
              schema:
                type: string
                example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
              required: true
            Set-Cookie:
              description: Кука для токена відновлення.
              schema:
                type: string
                example: __Host-4V5CcKvkhqt2pPYqDVhjpFFyswj5dqpG=eyJraWQiOiI4YzM4MDViMS1mZWU2LTQ4NmEtYmZlNi05M2QzYWU3YzQ2ZGYiLCJlbmMiOiJBMTkyR0NNIiwiYWxnIjoiZGlyIn0..aQR0XLwALU5Ek-8e.X85s6X6caFvzYQWirARGyoi8TkQPqlSIlLmQlM0HgoIVL7HW9e7QLsLjTBOZ5Oa5wKqh2acdc04b66vTInXzObh8iw0x3tXwmYXUAEi0eSTygH6SUvcRGOjQ_53xr9K_bkn-1rIPyfeVej1A.RG3-TAkpHwBCYMUfwuETsQ; Max-Age=604799; Expires=Sun, 17 Aug 2025 16:42:10 GMT; Path=/; Secure; HttpOnly
              required: true
        "400":
          description: Користувач помилився в наданні інформації.
          content:
            application/json:
              schema:
                oneOf:
                  - type: string
                  - $ref: '#/components/schemas/ErrorPayload'
        "403":
          description: Доступ не надано.
        "500":
          description: Невідома помилка.
  /authentication/login:
    put:
      summary: Авторизація користувача.
      operationId: login
      tags:
        - Authentication Controller
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginPayload'
      responses:
        "200":
          description: Користувача успісно авторизовано.
          headers:
            Authorization:
              description: Токен доступу.
              schema:
                type: string
                example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
              required: true
            Set-Cookie:
              description: Кука для токена відновлення.
              schema:
                type: string
                example: __Host-4V5CcKvkhqt2pPYqDVhjpFFyswj5dqpG=eyJraWQiOiI4YzM4MDViMS1mZWU2LTQ4NmEtYmZlNi05M2QzYWU3YzQ2ZGYiLCJlbmMiOiJBMTkyR0NNIiwiYWxnIjoiZGlyIn0..aQR0XLwALU5Ek-8e.X85s6X6caFvzYQWirARGyoi8TkQPqlSIlLmQlM0HgoIVL7HW9e7QLsLjTBOZ5Oa5wKqh2acdc04b66vTInXzObh8iw0x3tXwmYXUAEi0eSTygH6SUvcRGOjQ_53xr9K_bkn-1rIPyfeVej1A.RG3-TAkpHwBCYMUfwuETsQ; Max-Age=604799; Expires=Sun, 17 Aug 2025 16:42:10 GMT; Path=/; Secure; HttpOnly
              required: true
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetUserPayload'
        "400":
          description: Користувач помилився в наданні інформації.
          content:
            application/json:
              schema:
                oneOf:
                  - type: string
                  - $ref: '#/components/schemas/ErrorPayload'
        "403":
          description: Доступ не надано.
        "500":
          description: Невідома помилка.
  /authentication/get-me:
    get:
      tags:
        - Authentication Controller
      summary: Отримання авторизованого користувача.
      description: Отримати авторизованого користувача.
      responses:
        "200":
          description: Успішно.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetUserPayload'
        "500":
          description: Невідома помилка.
        "401":
          description: Неавторизовано.
          headers:
            WWW-Authenticate:
              required: true
              description: Спосіб аутентифікаці.
              schema:
                type: string
  /authentication/logout:
    delete:
      tags:
        - Authentication Controller
      summary: Вийти з акаунту.
      operationId: logout
      responses:
        "204":
          description: Успішно.
        "401":
          description: Неавторизовано.
          headers:
            WWW-Authenticate:
              required: true
              description: Спосіб аутентифікаці.
              schema:
                type: string